通过批量ID获取列表信息时，若列表的表单不存在，不返回该列表

+++ icity-iform-service/src/main/java/tech/ascs/icity/iform/controller/ListModelController.java
@@ -179,8 +179,12 @@ public class ListModelController implements tech.ascs.icity.iform.api.service.Li
 
 	@Override
 	public void removeListModel(@PathVariable(name="id") String id) {
-		listModelService.setItemReferenceListModelNull(id);
-		listModelService.deleteById(id);
+		ListModelEntity listModelEntity = listModelService.find(id);
+		if (listModelEntity!=null) {
+			listModelService.setItemReferenceListModelNull(id);
+			listModelService.deleteById(id);
+			listModelService.deleteListBtnPermission(id);
+		}
 	}
 
 	@Override
